Add authorization check to privileged endpoints (#290)
This PR adds authentication checks to some privileged endpoints (i.e.
cache hooks and cache flushes). To accomplish this task, [the Nest.js
recommended approach](https://docs.nestjs.com/security/authentication)
was followed. This includes the addition of the `passport` library,
which is recommended by Nest.js docs, and provides a pluggable and
strategy pattern-based approach to authentication management.

I've used a custom strategy (`passport-custom`) since I'm assuming we
want to keep compatibility with the Rust implementation (which expects a
`Basic <token>` header).

One of the advantages this strategy-based approach gives is that we
could use different strategies for different parts of the service API
(JWT for example), without too much effort (just plugging an additional
strategy subclass and a Nest.js guard).