Update external_request logs (#358)
To improve compatibility with Datadog defaults, some properties of the
`external_request` log changed. More specifically:

- `host` – is used as the originating host of the log by Datadog. We
want to represent to which host was the external request made to
instead. Changed to `target_host`.
- `status` – is used to set the level/severity of the log. We want to
represent the HTTP status code of the external request made instead.
Change to `request_status`
- `message` – its contents are displayed in views like Live Tail.
Changed to `detail` as `message` is already set as the JSON payload.

See
https://docs.datadoghq.com/logs/log_configuration/attributes_naming_convention/#reserved-attributes