Fix TransactionData addressInfoIndex property differences (#381)
Closes
https://github.com/safe-global/safe-client-gateway-nest/issues/362 and
https://github.com/safe-global/safe-client-gateway-nest/issues/367

This PR:
- Fixes the legacy Rust implementation differences by avoiding
traversing the `dataDecoded` with an arbitrary depth size. This fix
makes the service only get results from the first level (as the Rust
implementation does).
- For the items contained in `addressInfoIndex`, the associated
`AddressInfo` is retrieved from `Token`, and from `Contract` as a
fallback. Then, `AddressInfo` with no `name` property are filtered out.
(to meet the legacy behavior)
- Unifies the `txData.addressInfoIndex` response as a non-empty object,
or `null`.