Do not cache by default
If no expiration time is provided to the Data Source layer, do not cache the response by default.

We should explicitly declare the expiration time when needed and not assume that every request should be cached.

Additionally, skip E2E tests which assumed that the results were cached (these tests should be reviewed case by case).