Revert "Make AppModule a Dynamic Module (#794)" (#820)
This reverts commit 501bc743

The .env file is loaded with the ConfigModule. If we set an environment variable in .env then its value will not be part of the configuration (unless the environment variable is loaded in the current environment already).

This means that that the configuration that was loaded in the dynamic AppModule might have not taken into account the setup done in the .env file.

Since this was the main reason for making the AppModule dynamic, it is being reverted back to a standard module.