Add Emails Datasource (#829)
Datasource

Adds a `IEmailDataSource` – this component can be used to interact with the underlying data source which stores the email addresses for a given signer, for a given safe address on a specific chain. The following functionality was added:
  * `saveEmail` – saves an email address in the database. Returns a verification code if it was successfully stored.
  * `setVerificationCode` – resets the verification code for a signer. Returns the new verification code if successful.
  * `verifyEmail` – verifies a specific email given the code that was returned on registration. If the code matches then the email is considered verified. There's a configurable expiration time for the code that was introduced for security reasons – the code needs to be verified within a given timeframe to be considered valid. If the code is not verified in this timeframe then it should be reset.

Database

- Adds one table – `signer_emails` under the `emails` schema.
- A unique constraint exists for entries on `signer_emails`: an entry is considered unique when the following fields are considered: `chain_id`, `safe_address`, `signer`.

Migration Support

Migration support was added. Migration files need to be under the `migrations` folder. Inside it, a migration folder should be created with 5 digits and under that folder, an `index.sql` containing the migration file. 

For example, the initial migration has the following structure:

migrations/
└── 00001_initial
    └── index.sql

The initial migration creates the `signer_emails` table. The migration is done whenever the `PostgresDatabaseModule` is initialised (via the `PostgresDatabaseMigrationHook`).

Note on running multiple instances of the service: an advisory lock is acquired so that only one instance performs the migration. Other instances will wait for the lock to be released (via `pg_advisory_lock`). Migrations should be idempotent so that the final state of the database is the same upon re-execution of the migration file. See https://postgresql.org/docs/16/interactive/explicit-locking.html#ADVISORY-LOCKS

Docker compose support

- Adds a `db-test` container to Docker Compose – this way, given a development environment, two databases can be launched: the `db` one with an explicit mapped volume for development  (`/data` folder) and the `db-test` with no explicit mapped volumes which can be used for testing.
- The `db-test` container has different environment variable to setup its own location and default user. By default `db` will be exposed on port `5432` of the host while `db-test` will be exposed on port `5433`.