Add OnlySafeOwner guard
The `OnlySafeOwner` is a guard is a guard that can be used to validate payloads where a signature is provided.

This guard is only successful if the provided signature is valid and the provided signer is an owner of the safe.

This guard does not validate the contents of the message. Any security consideration (e.g. to avoid replay attacks) should be considered by the routes using this guard.

The guard can be applied to any route that has:
- the 'chainId' declared as a parameter
- the 'safeAddress' declared as a parameter
- the 'signature' as part of the JSON body (top level)
- the 'message' as part of the JSON body (top level)
- the 'address' as part of the JSON body (top level)

Additionally, the `ISafeRepository` now has a function `isOwner` which can be used to verify if an address is an owner of the safe.