Refactor date label timezone adjustment (#925)
Whenever a timezone offset was provided, we would adjust the underlying date directly with `setUTCMilliseconds`.

`setUTCMilliseconds` adjusts the milliseconds in the Date object and expects a value between 0 and 999. However, in the context of timezones, this value is much greater than 999 which is also addressed by the same function by adjusting the other time components (seconds, minutes, etc.). See https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setUTCMilliseconds#description

This refactor makes the change clearer by adding/subtracting the amount of millisecond to the date itself. Previously, a condition around the zero value was required so that the number of milliseconds would not reset.

Additionally the function was changing the `Date` object directly which can affect other parts of the code which would use the same object. To solve this, `structuredClone` is used to to a deep clone of the Date object and change that copy instead.