fix: retrieve Safe address via recovery module address (#961)
This adjusts the recovery alert logic to retrieve the Safe address via the address of the module. The Transaction Service is queried by the module address whcih returns a list of all Safes which the module is enabled on. As recovery modules are deployed per Safe, this uses the first address in the list.

* fix: retrieve Safe address via recovery module

* fix: lint

* fix: lint