Access schema service with 'this' reference
Accesses the `jsonSchemaService` with the instance reference (`this`) instead of the local one.

The `jsonSchemaService` instance is the same, but by accessing it with `this` we are more consistent on how we access the instance properties.

This also prevents the code from compiling if the property modifiers are removed (`private readonly`).