Add email subscriptions datasource (#1028)
Adds two new tables: `subscriptions` and `account_subscriptions`.

The `subscriptions` table holds a collection of subscription types:

|         id         |         key          |     name      |
|:------------------:|:--------------------:|:-------------:|
| SERIAL PRIMARY KEY | TEXT NOT NULL UNIQUE | TEXT NOT NULL |

The `key` should be used to reference the category in a human-readable format. The `name` should be used for rendering purposes.

A default entry is added with `migrations/000001_initial`: `key='account_recovery' name='Account Recovery'`. (Note: migration was edited directly because there's no deployment yet).

The `account_subscriptions` table holds the data on the account subscriptions which are active (i.e., if there is a row, then there is a subscription).

| account_id | subscription_id |
|:----------:|:---------------:|
| FOREIGN KEY (emails.account_emails) | FOREIGN KEY (emails.subscriptions) |

---
The following functionality was added to `IEmailDataSource`:
- `getSubscriptions` – Gets all the subscriptions for the account on chain id, with the specified Safe address.
- `subscribe` – Subscribes the account on chain id, with the Safe address to a category.
- `unsubscribe` – Unsubscribes from the notification category with the provided category key.
- `unsubscribeAll` – Unsubscribes from all notification categories.

---------

Co-authored-by: Aaron Cook <aaron@safe.global>
Co-authored-by: Héctor Gómez <hector.gomez.varela@gmail.com>