Wait for redis connection
When creating a new connection with `redisClientFactory`, we do not wait for the connection to be established and return the promise immediately.

While the promise will end up being resolved, it can prove more difficult to debug issues if there's an attempt to connect to the client before the promise is resolved.