Add email subscription cancellation routes (#1037)
- Adds `DELETE /v1/subscriptions?category=<string>&token=<string>`
  * This endpoint deletes the subscription of the given `category`, if a valid token is provided.
  * The token must be a UUID. A `400` is returned otherwise.
- Adds `DELETE /v1/subscriptions/all?token=<string>`
  * This endpoint deletes all subscriptions of an account with the associated `token`.
  * The token must be a UUID. A `400` is returned otherwise.
- Saving a new email with `EmailRepository` now subscribes to the Account Recovery category – this is a default subscription for all new accounts registered.

An indication of the token's existence is not returned to the client i.e. if an invalid token is provided the response is still successful. This was done while considering that the endpoint is unauthenticated and the link should be _private_ to the email owner. Therefore, no feedback is given regarding the validity of said token.