Return 204 for deletion of non-existent accounts (#1098)
If an account does not exist, we should return `204` if a deletion is triggered. The result of this action can be seen as idempotent as multiple consecutive deletions on the same account should not change the underlying state of the database.

Additionally, from a security standpoint, this also avoids leaking the existence of an account (although a signature is still required to trigger this endpoint).