Extract Verification Code table (#1101)
The verification code related data was extracted from the Accounts' table. Verification Codes are temporary data that should only be present while an account is being verified.

Previously, after an account was verified, most of the data tied to the verification process was set to null as it was still part of the accounts table. By extracting the verification code related data, we can now delete the respective rows instead of holding null values that were mostly used for temporary purposes.