Add getAccounts to AccountDataSource (#1130)
Adds a function to retrieve all the accounts tied to a Safe Address. This is particularly useful when only the Safe Address and chain id are available (e.g.: alert triggering).

The functionality was previously covered by `getVerifiedAccountEmailsBySafeAddress`. However, it was only returning the string values of the respective email addresses. While this covers some use-cases (to whom to send an email address), some other properties of the `Account` are still useful when sending email address such as the ability to check and send the unsubscription token.