Send unsubscription token when sending email
Sends a `unsubscriptionUrl` substitution to the email provider. This URL includes the token which can be used to unsubscribe from further email notifications.

Since this token is unique, and a token cannot be mapped to a target email address if multiple email addresses are involved, the `AlertsRepository` now triggers multiple calls to `IEmailApi.createMessage` (one per each recipient).