Replace jest.clearAllMocks with jest.resetAllMocks
The mock implementations were leaking in certain cases to other tests e.g.: if a specific test suite was missing a mock, adding the mock to a single test could've fixed the entire test suite.

By using `jest.resetAllMocks` we clear all the mocked implementations on top of the usage data on each mock.