Add UnauthenticatedExceptionFilter
The UnauthenticatedExceptionFilter can be used on routes which are not authenticated. The affected routes are:
- `POST /v1/chains/:chainId/safes/:safeAddress/emails/:signer/verify-resend` – it now returns `202 Accepted` under error conditions.
- `PUT /v1/chains/:chainId/safes/:safeAddress/emails/:signer/verify` – it now returns `400 Bad Request` under error conditions.

When used, this exception filter would catch any instance of `Error` and return to the clients the provided status code with an empty body.

This is specially useful for unauthenticated routes which need to provide feedback to the clients on a successful response while masking all errors under one response.