Use checksummed addresses in safeBuilder
Some routes, such as the Alert Controller routes, receive data from third parties which might include non-checksummed addresses.

The processing for some of these routes might include converting non-checksummed addresses to checksummed ones. If the Safe Builder doesn't provide checksummed addresses by default, we'd have to know if we should mock third party routes with a checksummed address or not.

With this new default, we provide support for mocking datasource calls that use checksummed addresses and non-checksummed ones (without having to change all the mocks that use the Safe address from the builder).