Enable Email Feature for Subscription Controller tests
The Subscription Controller tests were dependent on the value of the test configuration object, which means that if `features.email` was set to `false`, these tests would fail as the routes would not be installed.

By setting the value of `features.email` in the test setup, we isolate the test execution from the state of the test configuration object.