Limit maximum amount of TWAP parts (#1670)
Limits the number of parts we request up to 11 (confirmed by CoW) - which can be configured by an env. var. (`swaps.maxNumberOfParts`). If there are >11 parts, the final one is used to fetch the status and token information but it means that `executedSellAmount` and `executedBuyAmount` are `null` as we cannot calculate them:

- Add `swaps.maxNumberOfParts` configuration
- Update `TwapOrderInfo['executedSellAmount' | 'executedBuyAmount']` to be `string | null`
- Limit number of parts, fetching only the last if more than the maximum
- Add appropriate test coverage