fix: output the correct tx info type in swagger
Swagger is very picky. Our transactionInfoType is an enum and in every
TransactionInfo class we actually know the type of transactionInfo we
are dealing with. It’s impossible to have CreationTransactionInfo with
a “custom” type. The type is going to be “creation”.
By overriding the type with the correct type and setting an ApiProperty
with a single value the swagger docs manage to generate a type that
is closer to what the TransactionInfo is.