fix: swagger factory type can’t be null
The factory address is being set as AddressInfo | null. I think that it
in practice it is impossible for the factory to be null.

If we look at the transaction mapping in
src/routes/transactions/entities/transfer-transaction-info.entity.ts

We can see that the factory is set to
await this.addressInfoHelper.getOrDefault(…). The getOrDefault function
always returns an Address. More, the mapTransaction sets the transaction
as a CreationTransaction type. If we look at the zod schema:
export const CreationTransactionSchema = z.object({
…
  factoryAddress: AddressSchema,
…
});

So the zod defintion to the CreationTransaction is also an address.

Because of this I’ve changed the tx type generated by Swagger to be
only AddressInfo.