Disable Redis offline queue (#2304)
This PR adds the ability to disable the Redis offline queue to prevent messages from accumulating when the Redis server is down.

Currently, node-redis does not provide a way to configure the offline queue size. Since we use Redis solely as a cache server with a low TTL, there is almost no benefit in queuing messages while the server is unavailable.