fix: invite/role user org request bodies of Swagger (#2369)
The Swagger definitions for the request bodies of inviting users to `UserOrganization`s or updating their roles were missing. Both of theses were pure types and, as such, classes were created for them.

When importing an entity class, even if used as a `type`, it must be imported "normally" to be parsed by Swagger. This was the required fix for the `InviteUsersDto`:

- Create classes for both types
- Remove `type` from import