refactor: issue token from validated message (#2351)
We have to methods for Sign-In with Ethereum messages. This refactors the existing to use the simpler method implemented for org. onboard.

The `ISiweRepository` is now only responsible for parsing and validating a submitted `message`/`signature`. The `nonce` of which is compared against cached ones. As the upper limit for token expiration is auth-specific, that is now handled in the `AuthService`.

1. User calls `/nonce` and one is generated, cached and returned to the user.
2. User calls `/verify` and the message is parsed, validated, nonce compared, signature verified and then passed to the `AuthService`. The service ensures an `expirationTime` does not exceed the maxium and issues the token.

Note: on-chain signature verification was removed as it is not being used.

- Remove unused code from `ISiweRepository`.
- Move `expirationTime` check to `AuthService`.
- Add `iat` claim to token.
- Update tests accordingly.