fix: allow confirmation DTOs signatures under `signedSafeTxHash` property (#2423)
When adding [refinement of confirmations](https://github.com/safe-global/safe-client-gateway/pull/2382), we wrongly assumed only web used the confirmation endpoint. As such, we aligned the property name to match that of the Transaction Service.

Since then, we've confirmed that the property was correct. This allows both the new and previous name:

- Allow `signature` and `signedSafeTxHash` property, transforming the latter to `signature`
- Add appropriate test coverage