refactor: create `SafeSignature` entity (#2422)
For every confirmation refined, we reference the owner and type of a signature. This is the same for transactions and messages.

This moves shared logic into a new `SafeSignature` entity, [taking inspiration from `safe-eth-py`](https://github.com/safe-global/safe-eth-py/blob/main/safe_eth/safe/safe_signature.py).

It takes the `hash` signed and `signature`, and exposes the `r`, `s`, `v`, `signatureType` and `owner` properties of them:

- Remove existing signature helpers
- Create `SafeSignature` entity
- Use entity in place for transaction/message verification
- Add/update test coverage accordingly