refactor: centralise delegate call verification in verifier (#2427)
We refine confirmations and hashes of transaction proposal, confirmation and that from our API. We also block delegate calls by default or based on their trust. This was added as part of verification but it is not within the verifier.

This moves the relevant logic into the verifier, thereby centralising all logic in one place:

- Move delegate call verification into verifier
- Add/update test coverage appropriately