fix: ensure API signatures are even in length (#2434)
We currently validate that transaction signatures from our API are hex and a denomination fo 130 and that message signatures are hex.

Whilst we can be sure that incoming signatures from clients are a denomination of 130 (because the interface does not currently support proposal of contract signatures), we cannot be sure that the Transaction Service hasn't indexed contract signatures with dynamic parts.

This creates a new `HexBytesSchema` which ensures a string is hex and even in length, aka "byte-aligned", applying it to the validation of transaction/message signatures from our API:

- Add and use `HexBytesSchema`
- Add/update tests accordingly