fix: only throw for new `eth_sign` signatures (#2460)
Proposals with `eth_sign` throw. As it is still possible to submit an `eth_sign` signature directly, however, this means that concatenated signatures already present from our API would also cause errors.

This adds a new check to proposal verification, only throwing for `eth_sign` signatures if they are not already present.

An added benefit from this is also that existing signatures are also verified again:

- Fetch proposed transaction
- Only throw if an `eth_sign` signature is new
- Add/update test coverage