fix: remove excessive transaction verification checks (#2510)
On top of the transaction integrity/validity checks, we ensure there are no duplicate owners/signatures. As we inherently verify each signature, these checks are unnecessary. Their respective values would be validated. As we have also seen no logs for the above as well, this removes the respective duplicate checks (and the associated tests):

- Remove duplicate checks from `TransactionVerifier`
- Remove associated tests