refactor: migrate to Safe Decoder Service (#2501)
Migrates the domain/route layers to use the [Safe Decoder Service datasource](https://github.com/safe-global/safe-client-gateway/pull/2313). All existing usage of the data decoder of the Transaction Service has been ported to use the new Decoder Service.

Varying transactions fetched from the Transaction Service now use the new validation schema (which matches in structure to the previous one), with the addition of the new `accuracy` field. As the transactions do not have this, it will default to `UNKNOWN`.

The `data-decoder` controller now directs requests to the Decoder Service, requiring `to` and not accepting `value`:

- Update existing route-level `DataDecoded` entity to implement that of the Decoder Service
- Migrate existing entities to use the new `DataDecoded`:
  - Creation transactions
  - Module transactions
  - Multisig transactions
  - Data decoder
- Update test accordingly