fix: batch price requests (#2532)
Safes with a high number of assets are not able to retrieve prices. After looking into the issue, this is because the balance provider has a limit to the number of token addresses than can be requested at once.

This adjusts the way in which we requests prices: deduping token addresses and batching the number of requests according to the above limits:

- Add batch limit configuration
- Fetch prices in batches
- Add appropriate test coverage