feat: map batched `TokenInfo` into `tokenInfoIndex` (#2537)
Batched transfers are mapped client-side from the `dataDecoded`. As there is no associated `TokenInfo` for each transfer, it is not possible to parse the `value`, display the `name`, etc.

This adds a new `TransactionData['tokenInfoIndex']` property (similar to `addressInfoIndex`) which is populated with the `TokenInfo` of transferred assets.

The `dataDecoded` of a transfer is traversed, parsing the relevant token addresses interacted with and subsequently fetched for mapping into the aforementioned.

Note: this is preliminarily limited to 100 tokens but can be set dynamically according to a new `MAX_TOKEN_INFO` env. var.

- Add `mapping.transactionData.maxTokenInfoIndexSize` configuration.
- Add new `TransactionData['tokenInfoIndex']` property.
- Tweak types for parsing ERC-20 `dataDecoded`.
- Add new method to map batched `TokenInfo`.
- Add/update tests accordingly.