fix: use dynamic TTL for in-memory price cache (#2558)
We use a dynamic TTL when caching token prices - certain tokens have high refresh rates.

When populating the cache, this is [set correctly](https://github.com/safe-global/safe-client-gateway/blob/fix/in-memory-token-cache/src/datasources/balances-api/coingecko-api.service.ts#L376). When populating the in-memory cache, however, it is currently set to the defaul price TTL. This adjusts the logic in question to set the dynamic TTL:

- Populate in-memory cache according to token