fix: cache Kiln deployments relative to env. (#2570)
As vault deployments are only deployed on Base, we [dynamically use the testnet configuration of Kiln](https://github.com/safe-global/safe-client-gateway/pull/2526#discussion_r2031467964). This dynamic value is not taken into account when caching, however.

If the deployments are initially fetched on a network other than Base, the prod. deployments will be cached meaning that any subsequent deployment requests on Base are returned from the prod. cache and vice versa. This includes the dynamic URL in the cache key so as to distinguish the both.

- Include Kiln URL in cache key for deployments