fix: update amqplib dependency from 0.10.5 to 0.10.8 (#2581)
This PR resolves a connection error with the newest RabbitMQ version:
`negotiated frame_max = 4096 is lower than the minimum allowed value (8192)`
The previous version of amqplib hardcoded or defaulted to a `frame_max` of `4096`, which is below the minimum accepted by newer `RabbitMQ` versions. As a result, `AMQP` connection attempts failed during negotiation. This fix ensures compatibility with the server’s minimum frame size requirement.